Read input

+++ ledgerium_cleanup.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
 
-if [ -z "$1" ]
-  then
-    echo "Backup folder is not supplied"
-    exit 1
-fi
+# if [ -z "$1" ]
+#   then
+#     echo "Backup folder is not supplied"
+#     exit 1
+# fi
+
+# Timestamp function
+timestamp() {
+  date +"%Y_%m_%d_%H_%M_%S"
+}
+DIR="ledgeriumtools_$(timestamp)"
 
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
     
@@ -17,29 +23,32 @@ if [[ "$OSTYPE" == "linux-gnu" ]]; then
     [[ $DIST == *"centos"* ]]; then
         echo "OS: $DIST"
 
-        cd ../ledgeriumtools/output || exit
+        cd ..
+        if [ ! -d ledgeriumtools ]; then
+          echo "ledgerium ools doesnot exist, nothing to clean up"
+          exit
+        fi
+
+        cd ledgeriumtools || exit
         echo "Current folder - $PWD"
-
-        # Timestamp function
-        timestamp() {
-          date +"%Y_%m_%d_%H_%M_%S"
-        }
-
-        DIR=$1
-        echo "Backup folder $DIR"
-
-        echo 'Stopping all containers'
-        docker-compose down
-
-        echo 'Moving datastore files'
-        folder=output_"$(timestamp)"
-        mkdir -p "$DIR"/"$folder"
-        sudo mv -f tessera-* "$DIR"/"$folder"
-        sudo mv -f validator-* "$DIR"/"$folder"
-
-        echo 'Starting all containers'
-        docker-compose up -d
-    
+        
+        if [ ! -d $PWD/output ]; then
+          echo "ledgeriumtools backed up with $DIR"
+          cd ..
+          mv ledgeriumtools $DIR
+        else
+          cd output
+          if [ ! -f "docker-compose.yml" ]; then 
+            cd ../..
+            mv ledgeriumtools $DIR
+            echo "No running containers found, ledgeriumtools folder is backed up with $DIR"
+          else
+            docker-compose down
+            cd ../..
+            mv ledgeriumtools $DIR
+            echo "Existing containers are stopped and the existing ledgeriumtools folder is backed up with $DIR"
+          fi
+        fi
     else
         echo "Unknown linux distribution"
     fi
@@ -48,29 +57,30 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
         # Mac OSX
         echo "OS: $OSTYPE"
 
-        cd ../ledgeriumtools/output || exit
-        echo "Current folder - $PWD"
-
-        # Timestamp function
-        timestamp() {
-          date +"%Y_%m_%d_%H_%M_%S"
-        }
-
-        DIR=$1
-        echo "Backup folder $DIR"
-
-        echo 'Stopping all containers'
-        docker-compose down
-
-        echo 'Moving datastore files'
-        folder=output_"$(timestamp)"
-        echo "$DIR"/"$folder"
-        mkdir -p "$DIR"/"$folder"
-        sudo mv -f validator-* "$DIR"/"$folder"
-        sudo mv -f tessera-* "$DIR"/"$folder"
-
-        echo 'Starting all containers'
-        docker-compose up -d
+        cd ..
+        if [ ! -d ledgeriumtools ]; then
+          echo "ledgeriumtools doesnot exist, nothing to clean up"
+          exit
+        fi
+        cd ledgeriumtools || exit
+        
+        if [ ! -d $PWD/output ]; then
+          cd ..
+          mv ledgeriumtools $DIR
+          echo "ledgeriumtools backed up with $DIR"
+        else
+          cd output
+          if [ ! -f "docker-compose.yml" ]; then 
+            echo "No running containers found, ledgeriumtools folder is backed up with $DIR"
+            cd ../..
+            mv ledgeriumtools $DIR
+          else
+            docker-compose down
+            cd ../..
+            mv ledgeriumtools $DIR
+            echo "Existing containers are stopped and the existing ledgeriumtools folder is backed up with $DIR"
+          fi
+        fi
         
 elif [[ "$OSTYPE" == "cygwin" ]]; then
         # POSIX compatibility layer and Linux environment emulation for Windows
